+++ src/charactersData/index.ts
@@ -57,6 +57,8 @@ import ShikanoinHeizou from './characters/ShikanoinHeizou';
 import Tighnari from './characters/Tighnari';
 import Collei from './characters/Collei';
 import Dori from './characters/Dori';
+import Syno from './characters/Syno';
+import Candace from './characters/Candace';
 
 const CHARACTERS: ICharacter[] = [
   Amber,
@@ -115,7 +117,9 @@ const CHARACTERS: ICharacter[] = [
   ShikanoinHeizou,
   Tighnari,
   Collei,
-  Dori
+  Dori,
+  Syno,
+  Candace
 ]
 
 export default CHARACTERS