+++ CHANGELOG.md
@@ -1,4 +1,20 @@
-#### 1.1.12 September-2022
+#### 1.1.13 October-2022
+
+* Changes
+  * Added characters:
+    * Syno
+    * Candace
+  * Added boss:
+    * Algorithm of Semi-Intransient Matrix of Overseer Network
+  * Added materials:
+    * Scarab
+    * Redcrest
+    * Light guiding tetrahedron
+  * Added farming routes for:
+    * Scarab
+    * Redcrest
+
+* #### 1.1.12 September-2022
 
 * Changes
   * Added characters: